Restored the old csproj files and splitted up some class-files
