+++ Source/WhatTheHack/Harmony/DialogFormCaravanNotifyTransferablesChanged.cs
