+++ Source/WhatTheHack/Harmony/Need_MechEnergy_NeedInterval.cs
@@ -1,0 +1,13 @@
+﻿using HarmonyLib;
+using RimWorld;
+
+namespace WhatTheHack.Harmony;
+
+[HarmonyPatch(typeof(Need_MechEnergy), "NeedInterval")]
+internal static class Need_MechEnergy_NeedInterval
+{
+    private static bool Prefix(Need_MechEnergy __instance)
+    {
+        return !__instance.pawn.IsHacked();
+    }
+}
